pyproject.toml
[tool.poetry]
name = "sliderefactor"
version = "0.1.0"
description = "Convert NotebookLLM flattened slide PDFs into editable PPTX with SOTA accuracy"
authors = ["SlideRefactor Team"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-pptx = "^0.6.21"
opencv-python = "^4.8.1"
pillow = "^10.1.0"
requests = "^2.31.0"
pydantic = "^2.5.0"
paddleocr = {version = "^2.7.0", optional = true}
paddlepaddle = {version = "^2.5.2", optional = true}
jinja2 = "^3.1.2"
anthropic = "^0.20.0"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
pdf2image = "^1.16.3"
pypdfium2 = "^4.26.0"

[tool.poetry.extras]
ocr-fallback = ["paddleocr", "paddlepaddle"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"
mypy = "^1.7.1"
ruff = "^0.1.8"

[tool.poetry.scripts]
sliderefactor = "sliderefactor.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
